Do not subclass anymore the current pool implementation with Pool sub interfaces and instead use a proxy to a single pool implementation
